Add vpnaas_agent_package parameter to neutron::services::fwaas
When using both VPNaaS and FWaaS, puppet will remove
neutron-vpn-agent and install neutron-l3-agent instead.
This behavior is normal because neutron-vpn-agent is a superset
of neutron-l3-agent.

Adding vpnaas_agent_package as parameter to neutron::services::fwaas
to give the choice to install neutron-vpn-agent package instead of
neutron-l3-agent package on debian platforms.

Closes-Bug: 1328534